feat(settings): add deep links to specific sections (@fehmer) (#6661)
Allow to deep-link to  a specific config setting.

Users in discord often ask how to enable a specific setting. Instead of
guiding them manually to the correct settings we can use a deep-link:


https://github.com/user-attachments/assets/90b99660-13ae-49bf-8ff2-c5e6da290070

Added a link next to the group on mouse over to copy the deep link to
the clipboard:


![image](https://github.com/user-attachments/assets/d389754e-f43a-4638-b18f-729e818ef170)

---------

Co-authored-by: Miodec <jack@monkeytype.com>